requirements.txt
av==12.3.0
numpy>=1.24
pyzmq>=25.1
opencv-python-headless>=4.8
Pillow>=10.0
flask>=3.0
requests>=2.32
